tests/wpt/tests/css/css-masonry/tentative/parsing/masonry-slack-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masonry: masonry-slack getComputedStyle()</title>
<link rel="help" href="https://tabatkins.github.io/specs/css-masonry/">
<link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
<meta name="assert" content="masonry-slack computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<style>
  #target {
    font-size: 40px;
  }
</style>
<script>
test_computed_value("masonry-slack", "normal");

test_computed_value("masonry-slack", "10px");
test_computed_value("masonry-slack", "20%");
test_computed_value("masonry-slack", "calc(20% + 10px)");

test_computed_value("masonry-slack", "calc(-0.5em + 10px)", "0px");
test_computed_value("masonry-slack", "calc(0.5em + 10px)", "30px");
</script>
</body>
</html>
